Extract URL Beautifier interfaces
